Add a dreamup browser.
This adds a very simple dreamup browser. The browser is a web aplication with the frontend written in React, and the backend in Python on flask. The browser lets the user enter a url, then fetches the dreamup app contents from that url and generates html from it. There is also a mock nlp model server included for testing, so we don't have to run the (non-working) deep learning model to test out the browser.